// apps/demo/lib/content/types.ts
export interface NavLink {
	label: string
	href: string
	external?: boolean
	badge?: string
}